requirements.txt
# 3DRecon pipeline — Python dependencies
Pillow>=9.0
opencv-python>=4.5
numpy>=1.20
matplotlib>=3.5
tqdm>=4.60

# Optional: dock icon on macOS
# pyobjc-framework-Cocoa>=9.0
